pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-logging>=3.12.1",
    "pydantic-settings>=2.9.1",
    "streamlit>=1.42.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.7",
    "mypy>=1.15.0",
    "pytest>=8.3.2",
    "mdformat-gfm>=0.3.6",
    "mdformat-frontmatter>=2.0.8",
    "mdformat-footnote>=0.1.1",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint]

[tool.ruff.lint.isort]

[tool.mypy]
ignore_missing_imports = true
